Rename module data.local to data.local.database. Add module data.local.testing
