--- data/local/build.gradle
@@ -1,67 +1,0 @@
-import com.enricog.tempo.buildsrc.Libs
-import com.enricog.tempo.buildsrc.Versions
-
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'kotlin-kapt'
-}
-
-android {
-    namespace 'com.enricog.data.localdatasource'
-
-    compileSdkVersion Versions.compileSdk
-
-    defaultConfig {
-        minSdkVersion Versions.minSdk
-
-        javaCompileOptions {
-            annotationProcessorOptions {
-                arguments = [
-                        "room.schemaLocation": "$projectDir/schemas".toString(),
-                        "room.incremental"   : "true"
-                ]
-            }
-        }
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-
-
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-        }
-    }
-}
-
-
-dependencies {
-    implementation project(':data:routines:api')
-
-    implementation Libs.Kotlin.coroutines
-
-    // Room
-    implementation Libs.Room.runtime
-    kapt Libs.Room.compiler
-    implementation Libs.Room.ktx
-
-    // Hilt
-    implementation Libs.Hilt.android
-    kapt Libs.Hilt.androidCompiler
-
-    testImplementation project(':data:routines:testing')
-    testImplementation project(":core:coroutines:testing")
-    testImplementation Libs.Kotlin.Test.coroutines
-    testImplementation Libs.AndroidX.Test.core
-    testImplementation Libs.AndroidX.Test.rules
-    testImplementation Libs.AndroidX.Test.junitKtx
-    testImplementation Libs.AndroidX.Test.archCore
-    testImplementation Libs.Test.junit
-    testImplementation Libs.Robolectric
-    testImplementation Libs.Mockk.lib
-
-}
