+++ data/local/database/src/main/AndroidManifest.xml
