+++ data/local/database/src/main/java/com/enricog/data/local/database/routines/dao/SegmentDao.kt
@@ -1,0 +1,24 @@
+package com.enricog.data.local.database.routines.dao
+
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.Query
+import androidx.room.Update
+import com.enricog.data.local.database.routines.model.InternalSegment
+
+@Dao
+internal interface SegmentDao {
+
+    @Query("SELECT * FROM Segments ORDER BY rank DESC")
+    suspend fun getAll(): List<InternalSegment>
+
+    @Insert
+    suspend fun insert(vararg segment: InternalSegment): List<Long>
+
+    @Update
+    suspend fun update(vararg segment: InternalSegment)
+
+    @Delete
+    suspend fun delete(vararg segment: InternalSegment)
+}
