+++ data/local/database/src/main/java/com/enricog/data/local/database/routines/model/InternalRoutineWithSegments.kt
@@ -1,0 +1,18 @@
+package com.enricog.data.local.database.routines.model
+
+import androidx.room.Embedded
+import androidx.room.Relation
+import com.enricog.data.routines.api.entities.Routine
+
+internal data class InternalRoutineWithSegments(
+    @Embedded val routine: InternalRoutine,
+    @Relation(
+        parentColumn = "routineId",
+        entityColumn = "routineId_fk"
+    )
+    val segments: List<InternalSegment>
+) {
+    fun toEntity(): Routine {
+        return routine.toEntity(segments)
+    }
+}
